Split the UI class into Admin (responsible for registering admin hooks) and Settings (extend the WC_Settings_Page class)
